In H.264, check extradata[0], the config version
Some (all?) H.264 annexb files *do* generate an extradata field from
libav.js. Such files didn't decode properly, because the description
field would be set to this extradata, but it must not be set for annexb
files. I do not know whether this fix works in all cases.

At least in the example I have, extradata[0] is 0. extradata[0] is the
configuration version (always 1, but in the future certainly always not
0) in an AVCDecoderConfigurationRecord, so at least in this case, I can
use extradata[0] to distinguish between an annexb file and a non-annexb
file.

Fixes #2 (maybe)